Move classes to component package
