+++ app/src/main/java/com/example/jonathan/sectionsapp/component/BannerComponent.kt
@@ -1,6 +1,7 @@
-package com.example.jonathan.sectionsapp
+package com.example.jonathan.sectionsapp.component
 
 import android.view.View
+import com.example.jonathan.sectionsapp.R
 import com.example.jonathan.sectionsapp.databinding.BannerItemBinding
 
 class BannerComponent(
